[c][cpp] Apply skeleton.scaleX/scaleY to wind and gravity. See #2729.
Tested using the spine-sdl runtime and cloud-pot example for both spine-c and spine-cpp.

spine-c's PhysicsConstraint.c has the gravity inverted as it will need to be inverted if the runtime is y-down rather than y-up.
If the runtime is y-up, remove the negative sign and parentheses from self->gravity in line 143.